+++ weblate/urls.py
@@ -632,6 +632,26 @@ real_patterns = [
         name='memory-download',
     ),
     url(
+        r'^(?P<manage>manage)/memory/$',
+        weblate.memory.views.MemoryView.as_view(),
+        name='memory',
+    ),
+    url(
+        r'^(?P<manage>manage)/memory/delete/$',
+        weblate.memory.views.DeleteView.as_view(),
+        name='memory-delete',
+    ),
+    url(
+        r'^(?P<manage>manage)/memory/upload/$',
+        weblate.memory.views.UploadView.as_view(),
+        name='memory-upload',
+    ),
+    url(
+        r'^(?P<manage>manage)/memory/download/$',
+        weblate.memory.views.DownloadView.as_view(),
+        name='memory-download',
+    ),
+    url(
         r'^memory/project/' + PROJECT + '$',
         weblate.memory.views.MemoryView.as_view(),
         name='memory',